+++ src/main/java/insane96mcp/iguanatweaksreborn/mixin/AssignProfessionFromJobSiteMixin.java
@@ -1,0 +1,19 @@
+package insane96mcp.iguanatweaksreborn.mixin;
+
+import insane96mcp.iguanatweaksreborn.module.Modules;
+import net.minecraft.server.level.ServerLevel;
+import net.minecraft.world.entity.ai.behavior.AssignProfessionFromJobSite;
+import net.minecraft.world.entity.npc.Villager;
+import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.injection.At;
+import org.spongepowered.asm.mixin.injection.Inject;
+import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
+
+@Mixin(AssignProfessionFromJobSite.class)
+public class AssignProfessionFromJobSiteMixin {
+
+	@Inject(at = @At("TAIL"), method = "start(Lnet/minecraft/server/level/ServerLevel;Lnet/minecraft/world/entity/npc/Villager;J)V")
+	private void start(ServerLevel level, Villager villager, long gameTime, CallbackInfo callbackInfo) {
+		Modules.misc.villagerNerf.lockTrades(villager);
+	}
+}